pc-backend/backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.3
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cprofilev==1.4.0
cryptography==42.0.8
exceptiongroup==1.2.1
factory-boy==3.3.0
fakeredis==2.23.0
flask==3.0.3
flask-bcrypt==1.0.1
flask-cors==4.0.1
flask-jwt-extended==4.6.0
flask-limiter==3.6.0
flask-talisman==1.1.0
flower==2.0.1
frozenlist==1.4.1
gevent==24.2.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
kombu==5.4.0
locust==2.28.1
mako==1.3.5
markupsafe==2.1.5
marshmallow==3.21.3
msgpack==1.0.8
multidict==6.0.5
mypy==1.10.0
mypy-extensions==1.0.0
openpyxl==3.1.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
packaging==24.1
pip-tools==7.4.1
pluggy==1.5.0
prometheus-client==0.20.0
prometheus-flask-exporter==1.2.0
prompt-toolkit==3.0.47
protobuf==4.25.3
pycparser==2.22
pyinstrument==4.6.2
pyjwt==2.8.0
pyotp==2.9.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-flask==1.3.0
python-dateutil==2.9.0
python-json-logger==2.0.7
redis==5.0.6
requests==2.32.3
ruff==0.4.8
safety==3.2.3
sentry-sdk==2.5.1
six==1.16.0
sqlalchemy==2.0.30
swagger-ui-bundle==0.0.9
tomli==2.0.1
tornado==6.4
typing-extensions==4.12.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webauthn==2.1.0
webargs==8.4.0
werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zope-event==5.0
zope-interface==6.2